backend server/requirements.txt
fastapi
uvicorn
python-multipart
pillow
matplotlib
nltk
torch
tensorflow==2.5.0
tensorflow-serving-api==2.5.0
numpy
